chore: moved componenets to respective path folders
