+++ src/app/search/SearchAndFilterSection.tsx
